submit D.cpp to 'D - Static Sushi' (arc096) [C++14 (GCC 5.4.1)]

+++ 2018/0421_ARC096/D.cpp
@@ -36,7 +36,7 @@ typedef long long ll;
 
 int N;
 ll C;
-int x[100010];
+ll x[100010];
 ll v[100010];
 ll w[100010];
 ll w_rev[100010];
